+++ autoform/prob.py
@@ -21,6 +21,7 @@ import math
 from collections.abc import Hashable
 
 import autoform.ad as ad
+import autoform.constfold as constfold
 import autoform.core as core
 import autoform.dce as dce
 import autoform.memoize as memoize
@@ -30,6 +31,7 @@ import autoform.utils as utils
 __all__ = ["factor", "weighted"]
 
 factor_p = core.Prim("factor")
+constfold.constfold_rules[factor_p] = lambda eqn, _: eqn
 dce.non_dce_primitives.add(factor_p)
 memoize.non_memoizable_primitives.add(factor_p)
 
@@ -254,3 +256,10 @@ core.pull_bwd_rules.aset(weighted_call_p, utils.asyncify(pullback_bwd_weighted_c
 core.batch_rules.set(weighted_call_p, batch_weighted_call)
 core.batch_rules.aset(weighted_call_p, abatch_weighted_call)
 dce.dce_rules[weighted_call_p] = dce_weighted_call
+
+
+def constfold_weighted_call(eqn: core.Eqn, cond: constfold.ConstfoldCond, /) -> core.Eqn:
+    return eqn.using(ir=constfold.constfold(eqn.params["ir"], cond=cond))
+
+
+constfold.constfold_rules[weighted_call_p] = constfold_weighted_call